Branches decoded

+++ XM23p.h
@@ -181,7 +181,7 @@ void decodeInstructions();
 void printInstruction(int index, int wb, int rc, int src, int d, int flag);
 void printMoves(int index, unsigned char byte, int d);
 void printConCodes(int index, code strction);
-void printBranches(int index, int offset);
+void printBranches(int index, short offset);
 void printLdStr(int flag, int index, code strction);
 unsigned char concatByte(unsigned char b1, unsigned char b2);
 char concatLdStr(code strction);
